requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
openai
python-dotenv
requests
